+++ src/DGtal/geometry/volumes/BoundedRationalPolytope.h
@@ -771,7 +771,7 @@ namespace DGtal
 	    {
 	      Vector  n = ab.crossProduct( Point::base( k, (s == 0) ? 1 : -1 ) );
 	      Integer b = n.dot( pts[ i ] );
-	      int nb_in = 0;
+              std::size_t nb_in = 0;
 	      for ( auto p : pts ) {
 		Integer v = n.dot( p );
 		if ( v < b )  nb_in++;